# AsciiString optimized hashCode