… common/src/main/java/io/netty/util/internal/PlatformDependent0.java
… common/src/test/java/io/netty/util/internal/PlatformDependentTest.java