@Scottmitch 2015-11-04
Here are the benchmark results:

```
Benchmark                                          (size)   Mode  Cnt          Score         Error  Units
AsciiStringBenchmark.hashCodeBenchBytesNew              3  thrpt   10  171620686.411 ± 2067177.087  ops/s
AsciiStringBenchmark.hashCodeBenchBytesNew              5  thrpt   10  175861229.576 ± 3144324.972  ops/s
AsciiStringBenchmark.hashCodeBenchBytesNew              8  thrpt   10  160477033.757 ± 3303639.514  ops/s
AsciiStringBenchmark.hashCodeBenchBytesNew             10  thrpt   10  158626444.748 ± 1986456.633  ops/s
AsciiStringBenchmark.hashCodeBenchBytesNew             20  thrpt   10  127607141.124 ± 2167646.826  ops/s
AsciiStringBenchmark.hashCodeBenchBytesNew             50  thrpt   10   72125526.060 ± 2106200.452  ops/s
AsciiStringBenchmark.hashCodeBenchBytesOld              3  thrpt   10  176451138.408 ± 2384382.162  ops/s
AsciiStringBenchmark.hashCodeBenchBytesOld              5  thrpt   10  143466864.042 ± 1759612.341  ops/s
AsciiStringBenchmark.hashCodeBenchBytesOld              8  thrpt   10  109661110.010 ± 1154123.205  ops/s
AsciiStringBenchmark.hashCodeBenchBytesOld             10  thrpt   10  102299314.252 ± 1008274.904  ops/s
AsciiStringBenchmark.hashCodeBenchBytesOld             20  thrpt   10   62262391.840 ± 1098801.523  ops/s
AsciiStringBenchmark.hashCodeBenchBytesOld             50  thrpt   10   26230133.517 ±  357407.087  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew       3  thrpt   10  183354375.923 ± 3443497.773  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew       5  thrpt   10  158780695.388 ± 1622008.061  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew       8  thrpt   10  107521051.275 ± 3779937.365  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew      10  thrpt   10  100615689.343 ± 1529300.849  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew      20  thrpt   10   72518474.366 ± 1715818.963  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceNew      50  thrpt   10   38357354.866 ±  440889.222  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld       3  thrpt   10  166837329.528 ± 5377966.980  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld       5  thrpt   10  132701651.724 ± 2552045.916  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld       8  thrpt   10  101663377.948 ± 1089496.581  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld      10  thrpt   10   96768184.705 ± 1655143.710  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld      20  thrpt   10   59591234.529 ±  653538.861  ops/s
AsciiStringBenchmark.hashCodeBenchCharSequenceOld      50  thrpt   10   25853228.738 ±  384579.786  ops/s
```