@Scottmitch 2015-11-04
Here is the result of a `wrk` benchmark:

before optimized hash code:
```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    15.91ms   22.58ms 874.22ms   85.96%
    Req/Sec   197.29k    30.08k  496.01k    78.03%
  374716668 requests in 2.00m, 50.60GB read
Requests/sec: 3120069.60
Transfer/sec:    431.45MB
```

with this PR:
```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency    15.00ms   21.30ms 846.37ms   86.23%
    Req/Sec   210.93k    35.51k  355.74k    71.98%
  400726492 requests in 2.00m, 54.11GB read
Requests/sec: 3336640.04
Transfer/sec:    461.40MB
```