@Scottmitch 2015-11-04
In terms of how evenly the algorithm distributes I ran a collision test with the new and old algorithms.  I grabbed a list of 58120 English ascii words offline, plus netty's HTTP and HTTP/2 pseudo headers names, and then some headers found in the wild.

Collisions with old hash code
```
19007732 : excellency tract
```

Collisions with new hash code
```
777808595 : eagles moles 
808232867 : duvets sweets 
804164183 : makers users 
-840554786 : haulers hitters 
1566404641 : moated united 
2082282164 : efface menace 
-1292817698 : hauling hitting 
-1150087593 : making using 
257675040 : laager tiger 
```

The hash code collision rate is higher with the new algorithm for this data set. However non of the header values collided so It may not matter (feedback/comments welcome).