@normanmaurer 2015-11-05
@Scottmitch just a few comments. Overall LGTM.