@Scottmitch 2015-11-08
I tweaked the algorithm a bit and updated the performance results. The collisions were due to not dispersing the "least significant bytes" of the `byte[]` enough. I added additional left rotate operations when dealing with the last remaining input to get better dispersion. Now there are no collisions in this test set.