@Scottmitch 2015-11-10
I verified this PR on little endian unsafe/safe and also big endian safe. The big endian was emulated  (qemu, debian 7, ppc 32bit) and didn't provide unsafe, but this may be good enough to move forward.  All code that is accounting for endianness explicitly is verified correct, there may just be some code in the unsafe implementation that is not accounting for endiannes when it should be.

@normanmaurer - I think we have been operating under the assumption "review the endinaness code, and if it is broken someone will let us know". This PR takes it a step further and partially verifies correctness...is this good enough to merge?