@Scottmitch 2015-11-10
Cherry-picked 4.1 (b4b7913) master (33b74e3)